alex/bemo-docs: rename APIs
