+++ packages/sync/api-report.md
@@ -25,10 +25,21 @@ export interface TLMultiplayerUserInfo {
 }
 
 // @public
-export function useMultiplayerSync(opts: UseMultiplayerSyncOptions & TLStoreSchemaOptions): RemoteTLStoreWithStatus;
+export function useSync(opts: useSyncOptions & TLStoreSchemaOptions): RemoteTLStoreWithStatus;
 
 // @public
-export interface UseMultiplayerSyncOptions {
+export function useSyncDemo(options: useSyncDemoOptions & TLStoreSchemaOptions): RemoteTLStoreWithStatus;
+
+// @public (undocumented)
+export interface useSyncDemoOptions {
+    // @internal (undocumented)
+    host?: string;
+    roomId: string;
+    userInfo?: Signal<TLMultiplayerUserInfo> | TLMultiplayerUserInfo;
+}
+
+// @public
+export interface useSyncOptions {
     assets: TLAssetStore;
     // @internal (undocumented)
     onEditorMount?: (editor: Editor) => void;
@@ -39,17 +50,6 @@ export interface UseMultiplayerSyncOptions {
         [key: string]: any;
     }): void;
     uri: string;
-    userInfo?: Signal<TLMultiplayerUserInfo> | TLMultiplayerUserInfo;
-}
-
-// @public
-export function useSyncDemo(options: useSyncDemoOptions & TLStoreSchemaOptions): RemoteTLStoreWithStatus;
-
-// @public (undocumented)
-export interface useSyncDemoOptions {
-    // @internal (undocumented)
-    host?: string;
-    roomId: string;
     userInfo?: Signal<TLMultiplayerUserInfo> | TLMultiplayerUserInfo;
 }
